Change OutputView to OutputPresenter and add OutputView-stub
