+++ web-ui/src/main/java/org/archcnl/ui/MainPresenter.java
@@ -37,7 +37,7 @@ import org.archcnl.ui.menudialog.events.QuickOutputViewAccessRequestedEvent;
 import org.archcnl.ui.menudialog.events.RunToolchainRequestedEvent;
 import org.archcnl.ui.menudialog.events.ShowCustomQueryRequestedEvent;
 import org.archcnl.ui.menudialog.events.ShowFreeTextQueryRequestedEvent;
-import org.archcnl.ui.outputview.OutputView;
+import org.archcnl.ui.outputview.OutputPresenter;
 import org.archcnl.ui.outputview.sidebar.events.InputViewRequestedEvent;
 
 @Tag("MainPresenter")
@@ -46,7 +46,7 @@ public class MainPresenter extends Component {
     private static final long serialVersionUID = -8850076288722393209L;
     private static final Logger LOG = LogManager.getLogger(MainPresenter.class);
     private final MainView view;
-    private final OutputView outputView;
+    private final OutputPresenter outputPresenter;
     private final InputPresenter inputPresenter;
     private final ProjectManager projectManager;
 
@@ -58,11 +58,11 @@ public class MainPresenter extends Component {
         inputPresenter.addListener(ConceptHierarchySwapRequestedEvent.class, this::handleEvent);
         inputPresenter.addListener(RelationHierarchySwapRequestedEvent.class, this::handleEvent);
 
-        outputView = new OutputView();
-        outputView.addListener(ConceptGridUpdateRequestedEvent.class, this::handleEvent);
-        outputView.addListener(RelationGridUpdateRequestedEvent.class, this::handleEvent);
-        outputView.addListener(ConceptHierarchySwapRequestedEvent.class, this::handleEvent);
-        outputView.addListener(RelationHierarchySwapRequestedEvent.class, this::handleEvent);
+        outputPresenter = new OutputPresenter();
+        outputPresenter.addListener(ConceptGridUpdateRequestedEvent.class, this::handleEvent);
+        outputPresenter.addListener(RelationGridUpdateRequestedEvent.class, this::handleEvent);
+        outputPresenter.addListener(ConceptHierarchySwapRequestedEvent.class, this::handleEvent);
+        outputPresenter.addListener(RelationHierarchySwapRequestedEvent.class, this::handleEvent);
 
         view = new MainView(inputPresenter.getView());
         addListeners();
@@ -111,7 +111,7 @@ public class MainPresenter extends Component {
         view.addListener(FooterOptionRequestedEvent.class, FooterOptionRequestedEvent::handleEvent);
 
         inputPresenter.addListener(OutputViewRequestedEvent.class, e -> selectPathForChecking());
-        outputView.addListener(
+        outputPresenter.addListener(
                 InputViewRequestedEvent.class, e -> view.showContent(inputPresenter.getView()));
     }
 
@@ -129,15 +129,15 @@ public class MainPresenter extends Component {
     public void showOutputView(Optional<String> path) {
         try {
             ArchitectureCheck architectureCheck = new ArchitectureCheck();
-            outputView.setResultRepository(architectureCheck.getRepository());
+            outputPresenter.setResultRepository(architectureCheck.getRepository());
             if (path.isPresent()) {
                 runArchCnlToolchain(architectureCheck, path.get());
             }
-            outputView.displayResult(
+            outputPresenter.displayResult(
                     architectureCheck
                             .getRepository()
                             .executeNativeSelectQuery(QueryUtils.getDefaultQuery()));
-            view.showContent(outputView);
+            view.showContent(outputPresenter);
         } catch (PropertyNotFoundException e2) {
             view.showErrorMessage("Failed to connect to stardog database.");
         }
@@ -167,16 +167,16 @@ public class MainPresenter extends Component {
                 OpenProjectDialog openProjectDialog = new OpenProjectDialog(projectManager);
                 openProjectDialog.addListener(
                         ShowFreeTextQueryRequestedEvent.class,
-                        e -> outputView.showFreeTextQuery(e.getQuery(), e.isDefaultQueryTab()));
+                        e -> outputPresenter.showFreeTextQuery(e.getQuery(), e.isDefaultQueryTab()));
                 openProjectDialog.addListener(
                         ShowCustomQueryRequestedEvent.class,
-                        e -> outputView.showCustomQuery(e.getQuery(), e.isDefaultQueryTab()));
+                        e -> outputPresenter.showCustomQuery(e.getQuery(), e.isDefaultQueryTab()));
                 openProjectDialog.open();
                 break;
             case SAVE:
                 try {
                     projectManager.saveProject(
-                            outputView.getCustomQueries(), outputView.getFreeTextQueries());
+                            outputPresenter.getCustomQueries(), outputPresenter.getFreeTextQueries());
                 } catch (final IOException e) {
                     new ConfirmDialog("Project file could not be written.").open();
                 }
@@ -185,8 +185,8 @@ public class MainPresenter extends Component {
                 SaveProjectDialog dialog =
                         new SaveProjectDialog(
                                 projectManager,
-                                outputView.getCustomQueries(),
-                                outputView.getFreeTextQueries());
+                                outputPresenter.getCustomQueries(),
+                                outputPresenter.getFreeTextQueries());
                 dialog.addListener(
                         ProjectSavedEvent.class, e -> view.setSaveProjectMenuItemEnabled(true));
                 dialog.open();
